Move the copy resources task to the validate goal
